finished adding the delete function and api endpoint changed the getusers function to handle single and multiple results
